+++ src/state.ts
@@ -1,9 +1,0 @@
-import * as vscode from 'vscode';
-
-export interface State {
-    commandCall: boolean;
-    log: vscode.OutputChannel;
-    extensionId: string;
-    extensionName: string;
-    workspaceRoot?: string;
-}