+++ CHANGELOG.md
@@ -7,3 +7,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.0]
 
 - Initial release
+
+## [1.1.0]
+
+- Nitro API support
+- Two new configurations
+  - nuxtDeclarationNavigator.api.hover.enable [true/false]
+  - nuxtDeclarationNavigator.api.functions [$fetch, useFetch]
+- README updated